+++ app/sim_test.go
@@ -351,8 +351,6 @@ func TestAppStateDeterminism(t *testing.T) {
 	config := simcli.NewConfigFromFlags()
 	config.InitialBlockHeight = 1
 	config.ExportParamsPath = ""
-	config.OnOperation = false   //nolint:staticcheck
-	config.AllInvariants = false //nolint:staticcheck
 	config.ChainID = SimAppChainID
 	config.BlockMaxGas = SimBlockMaxGas